// vue/src/types/themeTypes/ThemeType.ts
export type ThemeTypes = {
  name: string;
  dark: boolean;
  variables?: object;
  colors: {
    primary?: string;
    secondary?: string;
    info?: string;
    success?: string;
    warning?: string;
    error?: string;
    lightprimary?: string;
    lightsecondary?: string;
    lightsuccess?: string;
    lightinfo?: string;
    lighterror?: string;
    lightwarning?: string;
    darkprimary?: string;
    darksecondary?: string;
    darkinfo?: string;
    darksuccess?: string;
    darkwarning?: string;
    darkerror?: string;
    darkText?: string;
    lightText?: string;
    borderLight?: string;
    inputBorder?: string;
    containerBg?: string;
    surface?: string;
    background?: string;
    'on-surface-variant'?: string;
    facebook?: string;
    twitter?: string;
    linkedin?: string;
    gray100?: string;
    primary200?: string;
    secondary200?: string;
    warning200?: string;
  };
};